Parameterize accelerators configuration
Two new gem5 options are added to parameterize the configuration of accelerators
in the simulated architecture.

--acc_types specifies the accelerator types to be instantiated in the system.
String type in the form of `ACC[,ACC]' is accepted, such as `Deblur' or
`Deblur,Denoise'. Currently the names are required to match those in the
lcacc.py, but adding simple parsing logic for alias is not hard.

--num_accinstances denotes the number of copies of each accelerator type. Given
`Deblur,Denoise' in --acc_types and --num_accinstances=2, accelerators will be
configured as deblur0, deblur1, denoise0, denoise1, with each may or may not
consisting more than one sub-accelerators depending on the application.

With that, change of accelerator configurations no longer requires
recompilation.